Move test models to fixtures
