+++ tests/Feature/UuidTest.php
@@ -6,8 +6,8 @@ use PHPUnit_Framework_TestCase;
 use Illuminate\Events\Dispatcher;
 use Illuminate\Container\Container;
 use Illuminate\Database\Capsule\Manager;
-use Tests\Post;
-use Tests\UncastPost;
+use Tests\Fixtures\Post;
+use Tests\Fixtures\UncastPost;
 
 class UuidTest extends PHPUnit_Framework_TestCase
 {
