+++ tests/Fixtures/UncastPost.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tests\Fixtures;
+
+use Tests\Fixtures\Model;
+
+class UncastPost extends Model
+{
+    //
+}
