--- tests/UncastPost.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Tests;
-
-class UncastPost extends Model
-{
-    //
-}
